# app/static/swagger.yaml
openapi: 3.0.0
servers:
  - url: http://127.0.0.1:8000
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Log in into system
      requestBody:
        required: true
        description: New token
        content:
          application/json:
            schema:
              $ref: '#components/schemas/LogIn'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: Logged in as {current_user_.username}
                access_token: {access_token}
                refresh_token: {refresh_token}
        '401':
           description: Bad Input
           content:
             application/json:
               example:
                 message: Wrong password
  /auth/registration:
    post:
      tags:
        - Authentication
      summary: Registration into system
      requestBody:
        required: true
        description: New user
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Register'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: User {username} was created
                access_token: {access_token}
                refresh_token: {refresh_token}
                'groups': groups
        '400':
          description: Bad input
          content:
            application/json:
              example:
                message: User {username} already exists
        '500':
           description: Error while register
           content:
             application/json:
               example:
                 message: Something went wrong
  /auth/refresh:
    post:
      security:
        - bearerAuth: []
      tags:
        - Authentication
      summary: Refresh access token
      requestBody:
        required: false
        description: New access token
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Refresh'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                access_token: {access_token}
                'groups': groups
        '400':
           description: Error while refresh
           content:
             application/json:
               example:
                 message: User {username} doesn't exist
  /auth/change-password:
    patch:
      security:
        - bearerAuth: []
      tags:
        - Authentication
      summary: Change password
      requestBody:
        required: true
        description: New password
        content:
          application/json:
            schema:
              $ref: '#components/schemas/Change_password'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: Password changed successfully
        '400':
          description: Bad input
          content:
            application/json:
              example:
                message: Incorrect password
        '404':
          description: User Not Found
          content:
            application/json:
              example:
                message: User not found.
  /halls:
    get:
      security:
        - bearerAuth: []
      tags:
        - Halls
      summary: Returns all halls in the cinema
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#components/schemas/HallsOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - Halls
      summary: Create new hall
      requestBody:
        required: true
        description: New hall info
        content:
          application/json:
            schema:
              $ref: '#components/schemas/HallIn'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Input
          content:
            application/json:
              example:
                message: Please, specify "name" and "capacity".
  /halls/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Halls
      parameters:
        - name: id
          in: path
          description: id of the hall
          required: true
      summary: Get one hall by id
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallOut'
        '404':
          description: Hall Not Found
          content:
            application/json:
              example:
                message: Hall not found.
    patch:
      security:
        - bearerAuth: []
      tags:
        - Halls
      summary: Updates hall's info
      parameters:
        - name: id
          in: path
          description: id of the hall
          required: true
      requestBody:
        required: true
        description: Hall's updated info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallIn'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: Updated.
        '404':
          description: Hall Not Found
          content:
            application/json:
              example:
                message: Hall not found.
    delete:
      security:
        - bearerAuth: []
      tags:
        - Halls
      summary: Remove hall by id
      parameters:
        - name: id
          in: path
          description: id of the hall
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: Hall was successfully deleted.
        '404':
          description: Hall Not Found
          content:
            application/json:
              example:
                message: Hall not found.
  /films:
    get:
      security:
        - bearerAuth: []
      tags:
        - Films
      summary: Returns all films in the cinema
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#components/schemas/FilmsOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - Films
      summary: Create new film
      requestBody:
        required: true
        description: New film info
        content:
          application/json:
            schema:
              $ref: '#components/schemas/FilmIn'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Input
          content:
            application/json:
              example:
                message: Please, specify "name", "genre", "director", "image" and "rating".
  /films/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Films
      parameters:
        - name: id
          in: path
          description: id of the film
          required: true
      summary: Get one film by id
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmOut'
        '404':
          description: Film Not Found
          content:
            application/json:
              example:
                message: Film not found.
    patch:
      security:
        - bearerAuth: []
      tags:
        - Films
      summary: Updates film's info
      parameters:
        - name: id
          in: path
          description: id of the film
          required: true
      requestBody:
        required: true
        description: Film's updated info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmIn'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: Updated
        '404':
          description: Film Not Found
          content:
            application/json:
              example:
                message: Film not found.
    delete:
      security:
        - bearerAuth: []
      tags:
        - Films
      summary: Remove film by id
      parameters:
        - name: id
          in: path
          description: id of the film
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              example:
                message: Film was successfully deleted.
        '404':
          description: Film Not Found
          content:
            application/json:
              example:
                message: Film not found.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HallIn:
      type: object
      properties:
        name:
          type: string
          example: IMAX
        capacity:
          type: integer
          example: 75
    HallOut:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: IMAX
        capacity:
          type: integer
          example: 75
...
    FilmIn:
      type: object
      properties:
        name:
          type: string
          example: Spiderman
        genre:
          type: string
          example: superhero
        director:
          type: string
          example: Sam Raimi
        rating:
          type: float
          example: 9.5
        image:
          type: string
          example: https://pics.filmaffinity.com/Spider_Man_No_Way_Home-771942059-large.jpg
    FilmOut:
      type: object
      properties:
        id:
          type: integer
          example: 4
        name:
          type: string
          example: Spiderman
        genre:
          type: string
          example: superhero
        director:
          type: string
          example: Sam Raimi
        rating:
          type: float
          example: 9.5
        image:
          type: string
          example: https://pics.filmaffinity.com/Spider_Man_No_Way_Home-771942059-large.jpg
...